x64 funzionante

+++ hmailserver/source/Server/Common/Mime/Mime.cpp
@@ -1359,8 +1359,9 @@ namespace HM
       for (BodyList::const_iterator it=bodies_.begin(); it!=bodies_.end(); it++)
       {
          // If the initial body ends with \r\n, remove them. We add new ones below.
+// PP patch cast a int
          if (bodies_.begin() == it && output.size() >= 2 && 
-            output[output.size()-2] == '\r' && output[output.size()-1] == '\n')
+            output[(int)output.size()-2] == '\r' && output[(int)output.size()-1] == '\n')
          {
             output = output.Mid(0, output.GetLength() - 2);
          }
